+++ sv-parser-error/src/lib.rs
@@ -20,8 +20,6 @@ pub enum ErrorKind {
     Preprocess,
     #[fail(display = "Define argument not found: {}", _0)]
     DefineArgNotFound(String),
-    #[fail(display = "Define text not found: {}", _0)]
-    DefineTextNotFound(String),
     #[fail(display = "Define not found: {}", _0)]
     DefineNotFound(String),
 }